Improve readme header (#70)
Previously the logo, description and the repo heading were all competing of the users attention by being visually equally prominent, making it hard for the user to parse this information. Also some url params were off and the navigation was kind of "lost" as it was squeezed between the two competing parties.

This change makes the header a secondary element which helps the user to scan the document. Also the navigation links are fixed and added a few items.