Fix/first tap to talk (#147)
What

Fix first tap-to-talk on first press

To make this work reliably I needed to turn mouseDown handler (which performs client.initialize on mouse press) async. To prevent further problems down stream, mouseUp handler needs to wait for the async mouseDown handler to finish.
There's no way to tell if the permission has been already granted. As a proxy, I've selected the duration of await client.initialize, as this will take a long time (>1500ms) if user interaction with the permission prompt is required. In this case the first tap-to-talk attempt is cancelled.

Why

Provide a reliable UX for voice functions

Tests (ok)

browser-ui testbench
fashion-ecommerce: Normal react-ui use
smart-home demo: Interoperation with poweron parameter
speech-to-text demo: Interoperation with no tap-to-talk

Other changes

In an attempt to make react-client more reliable and symmetric to browser-client, I've exposed clientState, startContext and stopContext calls. Previously, startContext() and stopContext() were concealed in favour of toggleRecording()